requirements.txt
astroid==2.11.7
asttokens==2.0.8
backcall==0.2.0
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
dill==0.3.5.1
entrypoints==0.4
executing==0.10.0
fonttools==4.37.1
gevent==21.12.0
greenlet==1.1.3
ipykernel==6.15.1
ipython==8.4.0
isort==5.10.1
jedi==0.18.1
jupyter-core==4.11.1
jupyter_client==7.3.5
kiwisolver==1.4.4
lazy-object-proxy==1.7.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mccabe==0.7.0
nest-asyncio==1.5.5
numpy==1.23.2
packaging==21.3
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pylint==2.14.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
pyzmq==23.2.1
scipy==1.9.0
six==1.16.0
stack-data==0.4.0
tomli==2.0.1
tomlkit==0.11.4
tornado==6.2
traitlets==5.3.0
wcwidth==0.2.5
websocket-client==1.4.0
wrapt==1.14.1
zope.event==4.5.0
zope.interface==5.4.0
